cat_dog_classifier/requirements.txt
numpy
opencv-python
matplotlib
tensorflowW
scikit-learn
streamlit
